move ui_action to Tools and fix missing screenshot when vision=False
