+++ droidrun/tools/adb.py
@@ -37,30 +37,6 @@ class AdbTools(Tools):
         # Store all screenshots with timestamps
         self.screenshots: List[Dict[str, Any]] = []
 
-    @staticmethod
-    def ui_action(func):
-        """"
-        Decorator to capture screenshots and UI states for actions that modify the UI.
-        """
-        @wraps(func)
-        def wrapper(*args, **kwargs):
-            self = args[0]
-            result = func(*args, **kwargs)
-            
-            frame = sys._getframe(1)
-            caller_globals = frame.f_globals
-            
-            step_screenshots = caller_globals.get('step_screenshots')
-            step_ui_states = caller_globals.get('step_ui_states')
-            
-            if step_screenshots is not None:
-                step_screenshots.append(self.take_screenshot()[1])
-            if step_ui_states is not None:
-                step_ui_states.append(self.get_state())
-            
-            return result
-        return wrapper
-
     def _parse_content_provider_output(
         self, raw_output: str
     ) -> Optional[Dict[str, Any]]:
@@ -108,7 +84,7 @@ class AdbTools(Tools):
         except json.JSONDecodeError:
             return None
 
-    @ui_action
+    @Tools.ui_action
     def tap_by_index(self, index: int) -> str:
         """
         Tap on a UI element by its index.
@@ -250,7 +226,7 @@ class AdbTools(Tools):
         """
         return self.tap_by_index(index)
 
-    @ui_action
+    @Tools.ui_action
     def swipe(
         self, start_x: int, start_y: int, end_x: int, end_y: int, duration: float = 0.3
     ) -> bool:
@@ -280,7 +256,7 @@ class AdbTools(Tools):
             print(f"Error: {str(e)}")
             return False
 
-    @ui_action
+    @Tools.ui_action
     def drag(
         self, start_x: int, start_y: int, end_x: int, end_y: int, duration: float = 3
     ) -> bool:
@@ -309,7 +285,7 @@ class AdbTools(Tools):
             print(f"Error: {str(e)}")
             return False
 
-    @ui_action
+    @Tools.ui_action
     def input_text(self, text: str) -> str:
         """
         Input text on the device.
@@ -360,7 +336,7 @@ class AdbTools(Tools):
         except Exception as e:
             return f"Error sending text input: {str(e)}"
 
-    @ui_action
+    @Tools.ui_action
     def back(self) -> str:
         """
         Go back on the current view.
@@ -373,7 +349,7 @@ class AdbTools(Tools):
         except ValueError as e:
             return f"Error: {str(e)}"
 
-    @ui_action
+    @Tools.ui_action
     def press_key(self, keycode: int) -> str:
         """
         Press a key on the Android device.
@@ -403,7 +379,7 @@ class AdbTools(Tools):
         except ValueError as e:
             return f"Error: {str(e)}"
 
-    @ui_action
+    @Tools.ui_action
     def start_app(self, package: str, activity: str | None = None) -> str:
         """
         Start an app on the device.
@@ -499,7 +475,7 @@ class AdbTools(Tools):
         except ValueError as e:
             raise ValueError(f"Error listing packages: {str(e)}")
 
-    @ui_action
+    @Tools.ui_action
     def complete(self, success: bool, reason: str = ""):
         """
         Mark the task as finished.
